+++ engine/src/main/java/net/jqwik/engine/facades/RandomGeneratorFacadeImpl.java
@@ -32,12 +32,6 @@ public class RandomGeneratorFacadeImpl extends RandomGenerator.RandomGeneratorFa
 	}
 
 	@Override
-	@Deprecated
-	public <T> RandomGenerator<T> withSamples(RandomGenerator<T> self, T[] samples) {
-		return new WithSamplesGenerator<>(self, samples);
-	}
-
-	@Override
 	public <T> RandomGenerator<T> unique(RandomGenerator<T> self) {
 		return new UniqueGenerator<>(self);
 	}