+++ engine/src/test/java/net/jqwik/engine/TestHelper.java
@@ -110,11 +110,6 @@ public class TestHelper {
 			}
 
 			@Override
-			public PropertyLifecycleContext propertyContext() {
-				return null;
-			}
-
-			@Override
 			public String label() {
 				return null;
 			}